feat: Changed all fatabase tables to match laravel database. Added /user endpoint

+++ src/main/java/com/spring/spring_booking_system/configs/JwtAuthenticationFilter.java
@@ -44,7 +44,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         }
 
         final String jwt = authHeader.substring(7);
-        final Integer userId;
+        final Long userId;
 
         try {
             userId = jwtService.extractId(jwt);
